pyproject.toml
[project]
name = "atheweb"
version = "0.1.0"
description = "TBD"
requires-python = ">=3.11"
authors = [
    { name = "Evan Chen", email = "evan@evanchen.cc" },
    { name = "Violet Xu", email = "violet@vbnm.dev"},
    { name = "Greta Qu", email = "greta.qu@athemath.org" },
    { name = "Ritwin Narra", email = "ritwin@stanford.edu" },
    { name = "Vera Puah", email = "v4nillacl0uds@gmail.com" }
]
dependencies = [
    "django>=5.2,<6.0",
    "django-allauth[socialaccount]>=65.0,<66.0",
    "django-bootstrap5>=24.2,<25.0",
    "django-extensions>=3.2,<3.3",
    "django-hijack>=3.4.0",
    "django-markdownfield>=0.11.0",
    "dotenv>=0.9.9",
    "ipython",
    "pillow>=10.0.0",
    "requests>=2.32.0",
    "django-discordo>=0.1.0",
    "pymdown-extensions>=10.21",
    "icalendar>=6.0,<7.0",
]

[project.optional-dependencies]
prod = [
    "gunicorn>=23.0.0",
    "mysqlclient>=2.2.7",
]

[dependency-groups]
dev = [
    "ruff",
    "djlint",
    "pytest",
    "pytest-django",
    "pyright",
    "django-stubs",
    "django-stubs-ext",
    "prek",
    "werkzeug>=3.1.6",
    "pytest-xdist>=3.8.0",
    "copier>=9.11.0",
    "codespell>=2.4.2",
]

[tool.djlint]
profile = "django"
ignore = "H006,H021,H023,H030,H031"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "atheweb.settings"
python_files = ["test_*.py", "tests.py"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
exclude = ["**/migrations/*.py", "manage.py"]

[tool.ruff.lint.extend-per-file-ignores]
"atheweb/settings.py" = ["S105"]
"**/tests.py" = ["S101"]
"**/test_*.py" = ["S101"]

[tool.pyright]
enableTypeIgnoreComments = true
exclude = ["**/node_modules", "**/typings", "**/__pycache__", ".git", "venv", ".venv", "**/migrations", "**/test_*.py", "**/tests.py", "**/apps.py"]
pythonPlatform = "Linux"
pythonVersion = "3.13"
typeCheckingMode = "basic"
useLibraryCodeForTypes = true

reportAssertAlwaysTrue = "error"
reportCallInDefaultInitializer = "warning"
reportConstantRedefinition = "none"
reportDuplicateImport = "error"
reportFunctionMemberAccess = "none"
reportGeneralTypeIssues = "error"
reportImplicitStringConcatenation = "none"
reportImportCycles = "none"
reportIncompatibleMethodOverride = "none"
reportIncompatibleVariableOverride = "none"
reportIncompleteStub = "none"
reportInvalidStringEscapeSequence = "error"
reportInvalidTypeVarUse = "warning"
reportMissingImports = "error"
reportMissingModuleSource = "warning"
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportOptionalCall = "error"
reportOptionalContextManager = "error"
reportOptionalIterable = "error"
reportOptionalMemberAccess = "error"
reportOptionalOperand = "error"
reportOptionalSubscript = "error"
reportOverlappingOverload = "error"
reportPrivateUsage = "error"
reportPropertyTypeMismatch = "error"
reportSelfClsParameterName = "warning"
reportTypedDictNotRequiredAccess = "warning"
reportUnboundVariable = "error"
reportUndefinedVariable = "error"
reportUnknownArgumentType = "none"
reportUnknownMemberType = "none"
reportUnknownParameterType = "error"
reportUnknownVariableType = "none"
reportUnnecessaryCast = "error"
reportUnnecessaryComparison = "error"
reportUnnecessaryIsInstance = "error"
reportUntypedBaseClass = "none"
reportUntypedClassDecorator = "none"
reportUntypedFunctionDecorator = "none"
reportUntypedNamedTuple = "error"
reportUnusedCallResult = "none"
reportUnusedClass = "error"
reportUnusedCoroutine = "error"
reportUnusedFunction = "error"
reportUnusedImport = "warning"
reportUnusedVariable = "error"
reportWildcardImportFromLibrary = "warning"
strictDictionaryInference = false
strictListInference = false
strictParameterNoneValue = false
strictSetInference = false

[tool.codespell]
skip = "*/migrations/*.py,uv.lock,fixtures/staff-snapshot-fa25.json,fixtures/course-snapshot-fa25.json"
ignore-words-list = "Oly"
